AWS: Use CAM service account (CAC script)
Implemented changes in the AWS CAC provisioning script to use a CAM
service account credentials file instead of a connector token. Modified
the script so that it downloads the CAM credentials from the S3 bucket,
then uses the CAM credentials to retrieve a new connector token. If
the deployment is using encryption, the script will decrypt the
credentials file like other secrets first before using it.

Signed-off-by: Edwin-Pau <epau@teradici.com>